Add database initialization and item search functionality
A new feature has been added to initialize the database, which incorporates error handling. The initialization procedure also includes populating the database with seed data if it is empty. Additionally, search functionality has been developed, which includes a new request helper class, an API endpoint in the SearchController, and a created index in the database. Finally, sample data file auctions.json was added to the project.