Add AuctionServiceHttpClient and update GetAllAuctions method
A new AuctionServiceHttpClient class was added to the SearchService. The GetAllAuctions method in the AuctionsController of the AuctionService was updated to support date filtering and enhance performance using an Iqueryable, which will make the application scale better for larger amounts of data. The method also now uses AutoMapper's ProjectTo method to map data.